Data file
Created a table to store biographical information about Thrawn.

Gave the whole table a thicker grooved border and set the cell borders only for the top and bottom.

Linked to thrawn-profile-picture for one of the cells.

Created the "infoheading" class, which makes the text slightly larger.

Created the "infolabel" class, which boldens the text and aligns with with the top.

Created the "infospec" class, which makes the text slightly smaller.