Data file positioning
Cleared the data file of right floats like the intro paragraph and floated it to the right.

Removed bottom margins from the intro paragraph and top margins from the data file to bring them closer together.

Removed the top border for the data file so their borders appeared to be blending together.