Quote styles
Split each quote up into multiple different paragraphs, each with a different class.

Used the "quotationmark" class to make the quotation marks bigger without changing the line height.

Used the "quotelabel" class to make the text slightly smaller and give it an indent.

Used the "infospec" class to make the text even smaller.

Set the margins for all these classes.

Gave the main text relative positioning and moved it up to make the quotation marks appear to be on the same line.

Created the "bottomquote" class for the bottom quote and gave it much smaller right margins so it could stretch out underneath the intro paragraph.