Fun fact and behind the scenes colors
Put two divs around "funfact" and gave all three elements a border so that I could fill the center, something that I could not do with the previous double border.

Made the outer and inner borders dark green and the middle border light green.

Repeated this same process with "behindscenes", except with green.