Media query (990px-1279px)
Made the page title smaller so it could all fit in one line.

Pushed the bottom quote to the left side of the intro paragraph instead of wrapping around it.

Decreased the top and bottom padding on the "Back to Top" button to reflect the text being broken up into two lines.