Media query (768px-989px)
Hide the page title and its underline while revealing the hidden title and underline that were located at the beginning of the flow instead of after the floated elements.

Un-floated and centered the intro paragraph and data file.

Hid the recommended media and the fun fact and behind the scenes bubbles.

Made the quotes take up their entire line instead of being next to the intro paragraph.

Centered the "bodycon" class and set its flex-direction to column instead of row.

Revealed the hidden fun facts and behind the scenes bubbles and the recommended media that were located toward the end of the normal flow instead of being floated.

Created a div to surround and center the fun facts and behind the scenes bubbles.

Got rid of the width restriction for the recommended media so it can expand to fill all of its space.